% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drape.R
\name{drape}
\alias{drape}
\title{Drape points to surface}
\usage{
drape(collar, topo, digits = 3)
}
\arguments{
\item{collar}{2D point simple feature with a 'geom' column}

\item{topo}{3D polygon simple feature with a 'geom' column}

\item{digits}{number of decimal digits used for rounding of elevations}
}
\value{
a vector containing the draped elevations of the points
}
\description{
Gets the elevation of the points draped on a surface.
}
