% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exportPly}
\alias{exportPly}
\title{Export Surface to a PLY File}
\usage{
exportPly(
  user,
  hostname,
  dbname,
  schema,
  tableName,
  srid = "3125",
  port = "5432"
)
}
\arguments{
\item{user}{PostGIS user}

\item{hostname}{PostGIS server host name or IP address}

\item{dbname}{database containing the polygon soup}

\item{schema}{schema containing the polygon soup}

\item{tableName}{table containing the polygon soup}

\item{srid}{spatial reference identifier of the geometry}

\item{port}{port number to connect to at the server host, or socket file
name extension for Unix-domain connections}
}
\value{
file name of exported ply file
}
\description{
A polygon soup represented by a PostGIS table with a PolygonZ column 'geom'
  is exported into a Stanford Polygon File (*.ply).
}
