% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{idAssetGroup}
\alias{idAssetGroup}
\title{Identify asset group}
\usage{
idAssetGroup(x)
}
\arguments{
\item{x}{a character vector representing the equipment}
}
\value{
a character vector representing the asset group
}
\description{
Identifies the asset group of a given equipment (or entity) which is the
  recipient of the item withdrawn from warehouse.
  This data can be seen from warehouse issuance slips prior to
  implementation of SAP.
}
\details{
Below are the identified groups:
  \itemize{
  \item Contractors
    \describe{
    \item{4K}{4K Development Corporation}
    \item{CKDI}{Cagdianao Konstruct Development Incorporated}
    \item{HPK}{HPK Construction}
    \item{MIZPEH}{}
    \item{MTEL}{}
    \item{NBI}{Nickelbase Incorporated}
    \item{PACE}{PrinceAce Corporation}
    \item{POLARIS}{Polaris Construction & General Services}
    \item{SDMC}{Southernphil Development and Marketing Corporation (FITZ-SDMC)}
    \item{SKAFF}{SKAFF Construkt Incorporated}
    \item{SMCC}{Sumitomo-Mitsui Construction Company}
    \item{TACOPHIL}{}
    \item{TRGC}{Tango Romeo General Construction}
    }
  \item Equipment
    \describe{
    \item{AC}{air compressor}
    \item{AD}{articulated dump truck}
    \item{AT}{all-terrain vehicle}
    \item{CM}{concrete mixer}
    \item{CT}{bulldozer}
    \item{DM}{Drill Machine}
    \item{DT}{dump truck}
    \item{FC}{fast craft}
    \item{FL}{forklift}
    \item{FT}{fuel truck}
    \item{FM}{farm tractors}
    \item{FR}{fire truck}
    \item{FS}{flying squid}
    \item{GS}{power generator set}
    \item{GT}{garbage truck}
    \item{HDT}{HOWO dump truck}
    \item{LC}{landing craft tank}
    \item{LM}{lawn mower}
    \item{MC}{mobile crusher}
    \item{MD}{mini dump truck}
    \item{ML}{man lift}
    \item{MR}{motorcycle}
    \item{MS}{mobile screener}
    \item{MT}{maintenance truck}
    \item{MX}{mini tracked excavator}
    \item{MW}{mini wheeled pay loader}
    \item{PATROLBOAT}{}
    \item{PB}{pump boat}
    \item{PM}{prime mover truck}
    \item{RG}{road grader}
    \item{SB}{service bus}
    \item{ST}{service truck}
    \item{SV}{service car}
    \item{TL}{tower light}
    \item{TX}{tracked excavator}
    \item{VR}{vibrating compactor}
    \item{WL}{wheeled pay loader}
    \item{WM}{welding machine}
    \item{WP}{water pump}
    \item{WR}{wheeled crane}
    \item{WT}{water truck}
    \item{WX}{wheeled excavator}
    }
  \item Fixed assets
    \describe{
    \item{ADMIN}{Admin building}
    \item{EQ}{Engineer's Quarters}
    \item{CIP}{construction in-progress}
    }
  \item Other parties
    \describe{
    \item{LGU}{local government unit}
    \item{MKTG}{marketing}
    \item{NCIP}{National Commission on Indigenous Peoples}
    \item{SECURITY}{}
    }
  \item Unknown
    \itemize{
    \item ZZZ
    }
  }
}
