% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{plyBinToDB}
\alias{plyBinToDB}
\title{Write binary PLY to database}
\usage{
plyBinToDB(
  inPly,
  user,
  hostname,
  dbname,
  tableName,
  schema = "public",
  port = "5432",
  srid = "3125"
)
}
\arguments{
\item{inPly}{name of ply file to be translated}

\item{user}{PostGIS user}

\item{hostname}{PostGIS server host name or IP address}

\item{dbname}{database to write into}

\item{tableName}{table to write into. This will be overwritten if it exists!}

\item{schema}{schema to write into}

\item{port}{port number to connect to at the server host, or socket file
name extension for Unix-domain connections}

\item{srid}{spatial reference identifier}
}
\value{
void
}
\description{
Reads the binary PLY file and writes it to a PostGIS database via streaming.
The table will have the columns: \code{geom geometry(PolygonZ, srid)} and
\code{slope_angle double precision}.
}
