% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{plyBinToText}
\alias{plyBinToText}
\title{Translate Binary PLY File}
\usage{
plyBinToText(in_ply, out_ply = "text.ply")
}
\arguments{
\item{in_ply}{name of ply file to be translated}

\item{out_ply}{name of output ply file}
}
\value{
void
}
\description{
Translates a binary PLY file to ASCII format.
}
