% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{plyToDB}
\alias{plyToDB}
\title{Write PLY to database}
\usage{
plyToDB(
  ply,
  db_user,
  db_host,
  db_name,
  db_table,
  schema = "public",
  port = "5432",
  srid = "3125"
)
}
\arguments{
\item{ply}{name of ply file}

\item{db_user}{PostGIS user}

\item{db_host}{PostGIS server host name or IP address}

\item{db_name}{database to write into}

\item{db_table}{table to write into. This will be overwritten if it exists!}

\item{schema}{schema to write into}

\item{port}{port number to connect to at the server host, or socket file
name extension for Unix-domain connections}

\item{srid}{spatial reference identifier}
}
\value{
void
}
\description{
Reads a PLY file and writes it to a PostGIS database via streaming.
The table will have the columns: \code{id integer} and
\code{geom geometry(PolygonZ, srid)}.
}
