% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{predictVars}
\alias{predictVars}
\title{Interpolate}
\usage{
predictVars(modelT, sample, radius, vario)
}
\arguments{
\item{modelT}{data frame of prediction model}

\item{sample}{data frame of sample coordinates}

\item{radius}{numeric value representing buffer radius}

\item{vario}{List contaning variogram models}
}
\value{
a dataframe containing the following:
  \enumerate{
    \item interpolated grade of element
    \item interpolation variance
    \item interpolation variance using absolute value of kriging weight
  }
}
\description{
This is used in a spatial study for limonite face samples.
}
