% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql.R
\name{psql}
\alias{psql}
\title{Execute psql Query}
\usage{
psql(
  db_host,
  db_user,
  db_name,
  query,
  ignore.stdout = TRUE,
  use_single_quoute = TRUE
)
}
\arguments{
\item{db_host}{PostGIS server host name or IP address}

\item{db_user}{PostGIS user}

\item{db_name}{database}

\item{query}{a single query to be executed}

\item{ignore.stdout}{see \link{system}}

\item{use_single_quoute}{use single quotes to enclose query string. If false,
double quotes are used instead.}
}
\value{
NULL if success
}
\description{
Execute an PostgreSQL query using psql
}
