% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readDTM}
\alias{readDTM}
\title{Read Surpac DTM file}
\usage{
readDTM(dtmFile, srid = "")
}
\arguments{
\item{dtmFile}{file name}

\item{srid}{spatial reference identifier}
}
\value{
a data.frame with the following columns:
  \describe{
    \item{area_2d}{2D area of the polygon}
    \item{edge_length}{length of the longest edge}
    \item{slope_angle}{angle in radians between z-axis and the polygon normal}
    \item{polygon}{polygon in ewkt or wkt format}
  }
}
\description{
Reads a valid DTM file. It is assumed that both the DTM and String files
  are present in the same directory and have the same file name. Outputs
  a data.frame with each row representing a triangle of the DTM.
}
