% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readPly}
\alias{readPly}
\title{Read PLY File}
\usage{
readPly(ply, srid = "")
}
\arguments{
\item{ply}{file name}

\item{srid}{spatial reference identifier of the geometry if output is ewkt}
}
\value{
a data.frame with the following columns:
  \describe{
    \item{area_2d}{2D area of the polygon}
    \item{edge_length}{length of the longest edge}
    \item{slope_angle}{angle in radians between z-axis and the polygon normal}
    \item{polygon}{polygon in ewkt or wkt format}
  }
}
\description{
Reads the ply an covert it to a data frame.
}
