% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readPlyText}
\alias{readPlyText}
\title{Read PLY File}
\usage{
readPlyText(plyFile)
}
\arguments{
\item{plyFile}{file name}
}
\value{
a data.frame with the following columns:
  \describe{
    \item{polygon}{polygon in ewkt or wkt format}
    \item{edge_length}{length of the longest edge}
    \item{slope_angle}{angle in radians between z-axis and the polygon normal}
  }
}
\description{
Reads the ply file created in \link{updateInsituSurface}.
}
