% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{tmctools-package}
\alias{tmctools}
\alias{tmctools-package}
\title{Tools for processing various data in Taganito Mine}
\description{
Tools used for processing some data in Taganito Mine.
}
\author{
\strong{Maintainer}: Basil Eric Rabi \email{basil.rabi@tmc.nickelasia.com}

}
