% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{variogramLine}
\alias{variogramLine}
\title{Wrapper for gstat::variogramLine}
\usage{
variogramLine(object, dist_vector, ...)
}
\arguments{
\item{object}{variogram model for which we want semivariance function values}

\item{dist_vector}{numeric vector or matrix with distance values}

\item{...}{other arguments passed to gstat::variogramLine}
}
\value{
a data frame with distance and semi-variance columns or a distance
  matrix
}
\description{
Semivariance Values For a Given Variogram Model
}
