% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{writeDTM}
\alias{writeDTM}
\title{Write Surpac DTM file}
\usage{
writeDTM(
  dtm,
  db_user,
  db_host,
  db_name,
  db_table,
  schema = "staging",
  port = "5432",
  srid = "3125"
)
}
\arguments{
\item{dtm}{file name}

\item{db_user}{PostGIS user}

\item{db_host}{PostGIS server host name or IP address}

\item{db_name}{database to write into}

\item{db_table}{table to write into}

\item{schema}{schema to write into}

\item{port}{port number to connect to at the server host, or socket file
name extension for Unix-domain connections}

\item{srid}{spatial reference identifier}
}
\value{
number of rows uploaded to the table
}
\description{
Reads a valid DTM file and writes it into PostGIS. It is assumed that the
  PostGIS user has its pgpass set-up.
}
\details{
It is also assumed that a table with access to a sequence is set-up like
  the \code{sql} query below:
  \preformatted{
CREATE SEQUENCE sequence_name
INCREMENT 1
START 1
MINVALUE 1
MAXVALUE 9223372036854775807
CACHE 1;

CREATE TABLE table_name
(
  id integer NOT NULL DEFAULT nextval('sequence_name'::regclass),
  edge_length double precision,
  geom geometry(PolygonZ,3125),
  CONSTRAINT table_name_pkey PRIMARY KEY (id)
)}
}
