% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{exportPly}
\alias{exportPly}
\title{Export Surface to a PLY File}
\usage{
exportPly(
  db_user,
  db_host,
  db_name,
  schema,
  db_table,
  srid = "3125",
  port = "5432",
  digits = 3L
)
}
\arguments{
\item{db_user}{PostGIS user}

\item{db_host}{PostGIS server host name or IP address}

\item{db_name}{database containing the polygon soup}

\item{schema}{schema containing the polygon soup}

\item{db_table}{table containing the polygon soup}

\item{srid}{spatial reference identifier of the geometry}

\item{port}{port number to connect to at the server host, or socket file
name extension for Unix-domain connections}

\item{digits}{precision in number of decimal digits}
}
\value{
file name of exported ply file
}
\description{
A polygon soup represented by a PostGIS table with a PolygonZ column 'geom'
  is exported into a Stanford Polygon File (*.ply).
}
