% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{idAssetGroup}
\alias{idAssetGroup}
\title{Identify asset group}
\usage{
idAssetGroup(x)
}
\arguments{
\item{x}{a character vector representing the equipment}
}
\value{
a character vector representing the asset group
}
\description{
Identifies the asset group of a given equipment (or entity) which is the
  recipient of the item withdrawn from warehouse.
  This data can be seen from warehouse issuance slips.
}
\details{
Below are the identified groups:
  \itemize{
  \item Contractors
    \describe{
    \item{HPK}{HPK Construction}
    \item{NBC}{Nickel Base Corporation}
    \item{PACE}{PrinceAce Corporation}
    \item{SDMC}{Southernphil Development and Marketing Corporation (FITZ-SDMC)}
    \item{SMCC}{Sumitomo-Mitsui Construction Company}
    \item{TACOPHIL}{}
    \item{TRGC}{Tango Romeo General Construction}
    }
  \item Equipment
    \describe{
    \item{ADT}{articulated dump truck}
    \item{ATV}{all-terrain vehicle}
    \item{CM}{concrete mixer}
    \item{CT}{bulldozer}
    \item{DT}{dump truck}
    \item{FASTCRAFT}{}
    \item{FL}{fuel lorry}
    \item{FORKLIFT}{}
    \item{FS}{flying squid}
    \item{FT}{fire truck}
    \item{FTR}{farm tractors}
    \item{GS}{power generator set}
    \item{GT}{garbage truck}
    \item{HDT}{HOWO dump truck}
    \item{LM}{lawn mower}
    \item{MC}{motorcycle}
    \item{MDT}{mini dump truck}
    \item{ML}{man lift}
    \item{MOC}{mobile crusher}
    \item{MT}{maintenance truck}
    \item{MTX}{mini tracked excavator}
    \item{MWL}{mini wheeled pay loader}
    \item{PATROLBOAT}{}
    \item{PB}{pump boat}
    \item{PMT}{prime mover truck}
    \item{RG}{road grader}
    \item{SB}{service bus}
    \item{SP}{service pickup}
    \item{ST}{service truck}
    \item{TL}{tracked pay loader}
    \item{TWL}{tower light}
    \item{TX}{tracked excavator}
    \item{VC}{vibrating compactor}
    \item{WC}{wheeled crane}
    \item{WL}{wheeled pay loader}
    \item{WP}{water pump}
    \item{WTL}{water lorry}
    \item{WX}{wheeled excavator}
    \item{YBM}{Yoshida Boring Machine}
    }
  \item Fixed assets
    \describe{
    \item{ADMIN}{Admin building}
    \item{EQ}{Engineer's Quarters}
    \item{CIP}{construction in-progress}
    }
  \item Other parties
    \describe{
    \item{LGU}{local government unit}
    \item{MKTG}{marketing}
    \item{SECURITY}{}
    }
  \item Unknown
    \itemize{
    \item WC
    \item ZZZ
    }
  }
}
