% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readplybin.R
\name{plyBinToText}
\alias{plyBinToText}
\title{Translate Binary PLY File}
\usage{
plyBinToText(inPly, outPly = "text.ply")
}
\arguments{
\item{inPly}{name of ply file to be translated}

\item{outPly}{name of output ply file}
}
\description{
Translates a binary PLY file to ASCII format.
}
